#741 removed old tests
